requirements.txt
pandas>=1.3.0
openpyxl>=3.0.7
xlrd>=2.0.0
requests>=2.25.1
python-dotenv>=0.18.0